Add Delete Profile button, Add auto-load last profile, fix overlay bug, fix new profile bug
- Added a Delete Profile button: This button will delete the currently
SELECTED (not loaded) profile. Text goes red on first click, profile is
deleted on second click. Click Load or Save or select a different
profile to cancel.
- Added auto-load last profile: The last profile loaded is automatically
saved. When loading the program again it will automatically select the
last used profile if it is available.
- Fixed a bug where the Overlays would not instantly toggle
HighQuality/AntiAlias/Curved, instead only pushing these updates when
something else about the had Overlay changed. These are not used yet in
this program but it is a bug in the base HOTK Overlays.
- Fixed a bug in the last commit where adding a new profile would fail
to set the default background colors for that profile.

Two save files are now used: programSettings.gd which stores application
data, and savedProfiles.gd which just stores the profile data. Old
profile data is renamed to savedSettings.gd.bak, instead of just
deleting the file, in case it fails to load them properly.

This closes #6, and closes #8.